Max flow: Added max flow to library and fixed warnings

+++ library/graph_lib/max_flow/max_flow.h
@@ -10,7 +10,7 @@
  * Class that solves the max flow problem. Implementation based on
  * Dinics Algorithm
  */
-template<typename FLOW_T, typename SIZE_T = int>
+template<typename FLOW_T, typename SIZE_T = long>
 class Max_flow
 {
 public:
