+++ library/graph_lib/max_flow/max_flow.tcc
@@ -22,7 +22,7 @@ void Max_flow<FLOW_T, SIZE_T>::resetGraph()
 template<typename FLOW_T, typename SIZE_T>
  void Max_flow<FLOW_T, SIZE_T>::connect(SIZE_T n1, SIZE_T n2, FLOW_T max_flow)
 {
-    if (n1 < graph.size() && n2 < graph.size())
+    if (n1 < static_cast<SIZE_T>(graph.size()) && n2 < static_cast<SIZE_T>(graph.size()))
     {
 	Max_flow<FLOW_T, SIZE_T>::Edge e{n2, 0, max_flow, false};
